requirements.txt
Django==4.2.7
django-extensions==3.2.3
django-crispy-forms==2.0
crispy-bootstrap5==0.7
PyPDF2==3.0.1
sentence-transformers==2.2.2
transformers==4.50.0
torch==2.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
langdetect==1.0.9
mistralai==0.1.2
django-widget-tweaks==1.4.12
django-tables2==2.6.0
django-filter==23.3